Final prediction API updated: log option select added

+++ final-project/prediction-api/prediction-api.R
@@ -32,7 +32,7 @@ function(req){
 #' @param user The user name querying the API
 #' @post /predict
 #' @get /predict
-function(jsondata, model, user, req) {
+function(jsondata, model, user, savelog = TRUE, req) {
   # Get the data as a dataframe 
   data <- fromJSON(jsondata)
   print("JSON data received from file")
@@ -50,48 +50,51 @@ function(jsondata, model, user, req) {
     prediction <- NA
   }
   
-  # Form a log list
-  log.list <- list(user = user,
-                   endpoint="admissions", 
-                   user.agent = req$HTTP_USER_AGENT,
-                   req.method =  req$REQUEST_METHOD,
-                   remote.addr = req$REMOTE_ADDR, 
-                   timestamp = now(), 
-                   model = model, 
-                   payload = data,
-                   prediction = prediction)
-  
-  ## Save the log in S3
-  # Generate a filename
-  filename <- paste(Sys.Date(), "-", substring(digest(paste(as.character(log.list$timestamp),
-                                       req$HTTP_USER_AGENT,
-                                       req$REMOTE_ADDR, sep = "+"),
-                                   algo = "sha1"), 1, 15), ".json", sep = "")
-  print(paste("Filename is:", filename))
-  
-  # Save the JSON file locally
-  log.json <- toJSON(log.list)
-  write_json(log.json, filename)
-  
-  # Push the JSON file in S3 bucket with credentials saved in ENV VARS
-  result <- tryCatch({
-    put_object(file = filename, 
-               object = filename, 
-               bucket = BUCKET_NAME, 
-               session_token = NULL)
-  }, warning = function(cond) {
-    print("A warning has raised. Original message: ")
-    print(cond)
-  }, error = function(cond) {
-    print("An error has ocurred. Original message: ")
-    print(cond)
-    print("# Ignoring saving log in S3")
-  }, finally = {
-    print("Processed file for S3 completed")
-    # Clean up the local log file
-    file.remove(filename)
-  })
-  
+  if (savelog) {
+    # Form a log list
+    log.list <- list(user = user,
+                     endpoint="admissions", 
+                     user.agent = req$HTTP_USER_AGENT,
+                     req.method =  req$REQUEST_METHOD,
+                     remote.addr = req$REMOTE_ADDR, 
+                     timestamp = now(), 
+                     model = model, 
+                     payload = data,
+                     prediction = prediction)
+    
+    ## Save the log in S3
+    # Generate a filename
+    filename <- paste(Sys.Date(), "-", substring(digest(paste(as.character(log.list$timestamp),
+                                                              req$HTTP_USER_AGENT,
+                                                              req$REMOTE_ADDR, sep = "+"),
+                                                        algo = "sha1"), 1, 15), ".json", sep = "")
+    print(paste("Filename is:", filename))
+    
+    # Save the JSON file locally
+    log.json <- toJSON(log.list)
+    write_json(log.json, filename)
+    
+    # Push the JSON file in S3 bucket with credentials saved in ENV VARS
+    result <- tryCatch({
+      put_object(file = filename, 
+                 object = filename, 
+                 bucket = BUCKET_NAME, 
+                 session_token = NULL)
+    }, warning = function(cond) {
+      print("A warning has raised. Original message: ")
+      print(cond)
+    }, error = function(cond) {
+      print("An error has ocurred. Original message: ")
+      print(cond)
+      print("# Ignoring saving log in S3")
+    }, finally = {
+      print("Processed file for S3 completed")
+      # Clean up the local log file
+      file.remove(filename)
+    })  
+  } else {
+    print("# Ignoring log saving")
+  }
   
   # Set up the output list for prediction
   out <- list(prediction = prediction)
